Amend deploy.sh to work with concourse
Commented out the download and install of the CF cli commands as our
docker image contains these.
Chmod the deploy script so we no longer have to do so each time we
deploy.
Two env variables are no longer required, we also do not need to
download the cf cli as the concourse worker has this code.

Single: @matthewcullum-gds